requirements.txt
django-oscar
